Rename y-common module to y-core
